requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
matplotlib>=3.4.3
seaborn>=0.11.2
joblib>=1.0.1
plotly>=5.1.0
dash>=2.0.0
supabase>=1.0.3
python-dotenv>=0.19.0
python-dateutil>=2.8.2
requests>=2.28.0